Revert "desktop-shell: avoid alternating surface between outputs"
This reverts commit 4eea29151240eb95564f137e6e612f9d583d3daf.

There were some details and cases that I've missed when writing this
commit, resulting in some weird behaviors. Trying to cover all of them
became a nightmare, and the function got really hard to read.

So it's better to revert this commit and think about other possible
solutions for the issue.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>