libweston: Update paint nodes in weston_output_repaint
This is when they need to be up to date. And it makes it so that
view_ensure_paint_node() only does what the function name indicates.

Also, later when we tie damage tracking to paint nodes it will make
more sense to update them just in time for the output being repainted.

Signed-off-by: Derek Foreman <derek.foreman@collabora.com>